\name{ComRiskModel-package}
\alias{ComRiskModel-package}
\docType{package}
\title{
Fitting of Complementary risk models
}

\description{
Evaluates the PDF, CDF, QF, random numbers and MLEs of eight well-known probability distributions in connection with complementary G binomial, complementary G negative binomial and complementary G geomatric family of distributions. Moreover, some commonly used data sets from the fields of actuarial, reliability, and medical science are also provided.
}

\details{
\tabular{ll}{
Package: \tab ComRiskModel\cr
Type: \tab Package\cr
Version: \tab 0.1 \cr
Date: \tab 2023-05-7\cr
License: \tab GPL-2\cr
}
}

\author{
Muhammad Imran <imranshakoor84@yahoo.com> and M.H Tahir <mht@iub.edu.pk>.
}

\section{Maintainers}{
Muhammad Imran <imranshakoor84@yahoo.com>
}

\references{
 Tahir, M. H., & Cordeiro, G. M. (2016). Compounding of distributions: a survey and new generalized classes. Journal of Statistical Distributions and Applications, 3, 1-35. <doi.org/10.1186/s40488-016-0052-1>.
}
\keyword{Parameter induction}
\keyword{Compounding}
\keyword{Distribution theory}
\keyword{Generalization of distribution}
\keyword{G binomial family of distributions}
\keyword{G geomatric family of distributions}
\keyword{G negative binomial family of distributions}


